S3C-193 reorganize level-net code structure
Put all files into a common rpc directory
