+++ index.js
@@ -41,7 +41,7 @@ module.exports = {
         http: {
             server: require('./lib/network/http/server'),
         },
-        level: require('./lib/network/level-net'),
+        level: require('./lib/network/rpc/level-net'),
     },
     storage: {
         metadata: {
