+++ lib/storage/metadata/file/client.js
@@ -4,7 +4,7 @@ const assert = require('assert');
 const Logger = require('werelogs').Logger;
 const constants = require('../../../constants');
 
-const levelNet = require('../../../network/level-net');
+const levelNet = require('../../../network/rpc/level-net');
 
 class MetadataFileClient {
