+++ lib/storage/metadata/file/server.js
@@ -4,14 +4,14 @@ const fs = require('fs');
 const os = require('os');
 const assert = require('assert');
 const uuid = require('uuid');
+const level = require('level');
+const sublevel = require('level-sublevel');
 
 const Logger = require('werelogs').Logger;
+
 const constants = require('../../../constants');
 const storageUtils = require('../../utils');
-
-const level = require('level');
-const levelNet = require('../../../network/level-net');
-const sublevel = require('level-sublevel');
+const levelNet = require('../../../network/rpc/level-net');
 
 const ROOT_DB = 'rootDB';
